CNFCERT-57 Placeholder go files for the new test suites
