+++ run-cnf-suites.sh
@@ -40,6 +40,7 @@ for var in "$@"
 do
 	case "$var" in
 		diagnostic) FOCUS="diagnostic|$FOCUS";;
+		access-control) FOCUS="access-control|$FOCUS";;
 		generic) FOCUS="generic|$FOCUS";;
 		multus) FOCUS="multus|$FOCUS";;
 		operator) FOCUS="operator|$FOCUS";;
